+++ js/quotation_view.js
@@ -12,8 +12,8 @@ $(function() {
         $('.item_price').each(function(i, el) {
             $.item_quantity = $(el).parent().find('.item_quantity');
             $.item_sub_total = $(el).parent().find('.item_sub_total');
-            var item_quantity = (isNaN(parseInt($.item_quantity.html()))) ? 0 : parseInt($.item_quantity.html());
-            var item_price = (isNaN(parseInt($(el).html()))) ? 0 : parseInt($(el).html());
+            var item_quantity = (isNaN(parseFloat($.item_quantity.html()))) ? 0 : parseFloat($.item_quantity.html());
+            var item_price = (isNaN(parseFloat($(el).html()))) ? 0 : parseFloat($(el).html());
             var tmp_value = item_quantity * item_price;
             $.item_sub_total.html(tmp_value);
         });
